requirements.txt
pydantic==1.10.8
PyYAML==6.0
superagi-tools==1.0.6
duckduckgo-search==3.8.3
requests==2.31.0